+++ clang-tools-extra/clang-tidy/bsl/LambdaImplicitCaptureCheck.cpp
@@ -1,0 +1,55 @@
+//===--- LambdaImplicitCaptureCheck.cpp - clang-tidy ----------------------===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#include "LambdaImplicitCaptureCheck.h"
+#include "clang/AST/ASTContext.h"
+#include "clang/ASTMatchers/ASTMatchFinder.h"
+
+using namespace clang::ast_matchers;
+
+namespace clang {
+namespace tidy {
+namespace bsl {
+
+void LambdaImplicitCaptureCheck::registerMatchers(MatchFinder *Finder) {
+  Finder->addMatcher(lambdaExpr().bind("lambda"), this);
+}
+
+void LambdaImplicitCaptureCheck::check(const MatchFinder::MatchResult &Result) {
+  auto Lambda = Result.Nodes.getNodeAs<LambdaExpr>("lambda");
+  auto Mgr = Result.SourceManager;
+
+  auto CaptureLoc = Lambda->getIntroducerRange();
+  if (CaptureLoc.isInvalid())
+    return;
+
+  auto Begin = CaptureLoc.getBegin();
+  if (Begin.isInvalid() || Begin.isMacroID())
+    return;
+
+  if (Mgr->getFileID(Begin) != Mgr->getMainFileID())
+    return;
+
+  auto CapList = Lambda->implicit_captures();
+
+  for (const auto &Cap : CapList) {
+    auto Var = Cap.getCapturedVar();
+    if (!Var)
+      continue;
+
+    if (Var->getStorageDuration() != SD_Automatic)
+      continue;
+
+    auto Warn = "implicit capture of variable with automatic storage duration";
+    diag(Cap.getLocation(), Warn);
+  }
+}
+
+} // namespace bsl
+} // namespace tidy
+} // namespace clang