+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -75,6 +75,11 @@ New module
 
 New checks
 ^^^^^^^^^^
+- New :doc:`bsl-lambda-implicit-capture
+  <clang-tidy/checks/bsl-lambda-implicit-capture>` check.
+
+  Check for implicit captures of variables with automatic storage duration.
+
 - New :doc:`cppcoreguidelines-avoid-non-const-global-variables
   <clang-tidy/checks/cppcoreguidelines-avoid-non-const-global-variables>` check.
   Finds non-const global variables as described in check I.2 of C++ Core
@@ -144,7 +149,7 @@ New checks
 
   Finds ``cnd_wait``, ``cnd_timedwait``, ``wait``, ``wait_for``, or
   ``wait_until`` function calls when the function is not invoked from a loop
-  that checks whether a condition predicate holds or the function has a 
+  that checks whether a condition predicate holds or the function has a
   condition parameter.
 
 - New :doc:`bugprone-reserved-identifier